Move new content collection to more standard location
